api/v1/contacts/serializer.py
from rest_framework import serializers

from api.models.contact import Contact


class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = '__all__'


