requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cors==1.0.1
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
filelock==3.12.0
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
whitenoise==6.4.0
zope.event==4.6
zope.interface==6.0
